requirements.txt
Keras==2.0.2
numpy==1.12.1
python_resize_image==1.1.11
scipy==0.19.0
tensorflow==1.1.0
Pillow==4.1.1
